ipacm: Fix tautological-constant-out-of-range-compare warning
IPA_HDR_L2_MAX is defined outside the ipa_hdr_l2_type enum, which
causes this warning. Since that enum is defined in uapi headers,
work around it in ipacm instead of having to modify all kernels.

+++ ipacm/src/IPACM_LanToLan.cpp
@@ -282,7 +282,7 @@ void IPACM_LanToLan::handle_iface_up(ipacm_event_eth_bridge *data)
 			return;
 		}
 
-		if(data->p_iface->tx_prop->tx[0].hdr_l2_type == IPA_HDR_L2_NONE || data->p_iface->tx_prop->tx[0].hdr_l2_type == IPA_HDR_L2_MAX)
+		if(data->p_iface->tx_prop->tx[0].hdr_l2_type == IPA_HDR_L2_NONE || (int)data->p_iface->tx_prop->tx[0].hdr_l2_type == IPA_HDR_L2_MAX)
 		{
 			IPACMERR("Invalid l2 header type %s!\n", ipa_l2_hdr_type[data->p_iface->tx_prop->tx[0].hdr_l2_type]);
 			return;
